feat: create attendance type extension
changes:
- `AttendanceTypeExtensions` class added

closes #309
