+++ src/Domain/Extensions/EnumeratorExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
 using Domain.Caches;
-using Domain.Enumerators;
 
 namespace Domain.Extensions;
 
@@ -45,24 +44,4 @@ public static class EnumeratorExtensions
 	/// <returns>The <see cref="DisplayAttribute.ShortName"/> or the <paramref name="value"/> as string.</returns>
 	public static string GetShortName<T>(this T value) where T : struct, IComparable, IFormattable, IConvertible
 		=> DisplayAttributeCache<T>.GetShortName(value);
-
-	/// <summary>
-	/// Returns if the attendance type is working hours relevant.
-	/// </summary>
-	/// <param name="type">The attendance type enumerator to work with.</param>
-	/// <returns>
-	/// <see langword="true"/> if the attendance type is working hours relevant, otherwise <see langword="false"/>
-	/// </returns>
-	public static bool IsWorkingHoursRelevant(this AttendanceType type)
-		=> type switch
-		{
-			AttendanceType.WORKDAY => true,
-			AttendanceType.ABSENCE => true,
-			AttendanceType.BUISNESSTRIP => true,
-			AttendanceType.MOBILEWORKING => true,
-			AttendanceType.SHORTTIMEWORK => true,
-			AttendanceType.VACATIONBLOCK => true,
-			AttendanceType.PLANNEDVACATION => true,
-			_ => false,
-		};
 }
